moved dm indicators to adx pacakge
